added save log option, fixed log reparsing
The save button provides an option to save the log content to a text file.
The update log is proposed to be saved to a file called "update-modules.log".
The run log ist proposed to be saved to a file besides the wrp file.

Toggling debug output now reparses the log that has been printed so far, showing old debug messages or hiding them.